stop reflector only when it's installed
when reflector in not installed systemctl stop fails

+++ alis.sh
@@ -317,7 +317,9 @@ function prepare() {
 
 function configure_reflector() {
     if [ "$REFLECTOR" == "false" ]; then
-        systemctl stop reflector.service
+        if systemctl is-active --quiet reflector.service; then
+            systemctl stop reflector.service
+        fi
     fi
 }
